Added basic minmax algorithm with very simple heuristic to evaluate leaf nodes
This code is not very efficient, as there is quite a lot of checking for win condition (which is expensive right now). Furthermore, the heuristic for evaluating a position is very simple - assuming the code is even free of bugs, it just scores more points the more central pieces you have.