Fairy large refactor (performance benefits)
Instead of using a coordinate-based format for storing generated moves, we switch to a bitboard format internally (the `Move` type is kept the same). This allows us to store multiple moves in one `InternalMove` representation (saving space), and more importantly avoids conversions between coordinate and bitboard representations.

Other than that, BoardPosition was renamed BoardCoordinates (to better distinguish it from `rock::Position`), a separate internal bit_operations.h header was added, and an `extract_one_bit()` function was added for the common use case of extracting one bitboard from a bitboard containing multiple set bits.

The pure alpha-beta implementation of recommend_move was removed. It is now the goal to keep a simple minimax version and the latest version only, instead of every version side-by-side (which would present a significant maintenance burden).

A simple test of move recommendation speed was added to the tests (these performance tests really don't belong in the same place as unit tests, but for now it's ok).