Fixed some easy efficiency wins, removed standard minimax search
Standard minimax search was removed because it was annoying keeping it up to date. The idea was it would provide a reference for a simple search, but that utility wasn't valuable enough to have essentially duplicated code lying around.

Efficiency was mainly gained by not re-generating moves pointlessly when
checking if the game was over.