Prevent repeating analysis on moves from the transposition table
If the transposition table contains an entry, that move is processed straight
away, but the move is not removed from the generated list. This is because we
can save even generating the list at all if this is a cut-node. It also avoid
searching through the list for the found-move and erasing it (this may be
avoided by remembering where in the list the move was, but that introduces some
additional complexity.