Improvements to transposition table, two options
Transposition table was allocating a unique_ptr for each node, now the memory
is flat. There are currently two options, a custom table that is backed by a
std::vector that must account for collisions (it accepts collisions and leaves
it up to the user to overwrite or not), and a table based on Abseil's
flat_hash_map (hash collisions are dealt with internally and not exposed to the
user).

In both cases the Abseil hash function was found to be fairly good; I am unable
to come up with something better right now, and Zobrist hashing is too slow
without incrementable re-hashing (which has not yet been implemented).

Most scenarios saw a significant performance benefit from using either new
table (perhaps 20% or more in some cases), but the starting position did not.
The improvement there was more like 5% in the best case. Varying the table size
of the new table altered performance, and in the worst cases the analysis of
the starting position saw a performance regression. This despite saving on
unique_ptr allocations. Therefore I assume that the starting position is more
prone to hash collisions than other more random positions.

In general the custom table was around 5-10% faster than the Abseil-backed
version. Unsurprising, since the Abseil table must perform resizing as well as
some form of table traversal when there is a hash collision.