Restructured internals of project, splitting up algorithms.cpp
Also added a pointer to a transposition table as a member of the Searcher class to avoid having that hard-coded to a global variable.

Removed ability to turn of killer heuristic and transposition tables through macros. The conditional compilation of these features was proving a maintenance burden, and since they are both fairly well-proven by now, these macro switches were removed. Negascout can still be disabled though, and the diagnostics are not enabled by default.