Add ROCK_API macro to export symbols if dynamic linking is used
This is necessary on Windows, as all public symbols must be marked with `__declspec(export)`. On Linux as I understand it, it is good practice to make the default symbol visibility hidden, and then mark public symbols visible (this reduces load times of shared object files).

The `ROCK_API` macro is defined in `include/rock/common.h` in the same way as `fmtlib` defined `FMT_API` in their `core.h` file. The exporting of symbols is activated if compiling as a shared library; in the same way as `fmtlib` we defined `ROCK_EXPORT` from the CMakeLists.txt file to activate this behaviour if we are building as a shared library.