Properly evaluate draws in the search
If the game ends partway through one branch of the search, the engine was recording the score (win or lose) and returning it. But if the position was a draw then the normal evaluation would be returned. This meant the engine would not correctly value draw positions (which should always evaluate to 0).

This flaw is fixed in this commit.